refactor: AspNetCore solution
