+++ packages/scalar.aspnetcore/src/ScalarTargets.cs
@@ -58,22 +58,3 @@ public enum ScalarTargets
     [Description("swift")]
     Swift
 }
-
-internal static class Extensions
-{
-    public static string GetDescription(this Enum enumValue)
-    {
-        var field = enumValue.GetType().GetField(enumValue.ToString());
-        if (field == null)
-        {
-            return enumValue.ToString();
-        }
-
-        if (Attribute.GetCustomAttribute(field, typeof(DescriptionAttribute)) is DescriptionAttribute attribute)
-        {
-            return attribute.Description;
-        }
-
-        return enumValue.ToString();
-    }
-}
